Refactor posts list component to use PascalCase
Why:

* Following up on the work started in #362, but with a more limited
  scope to prevent breaking in progress work.

This change addresses the need by:

* Renaming the posts list component files to follow the Airbnb
  convention of naming the files after the same entity they export by
  default, which imposes using PascalCase for component files.
