+++ src/components/blog/PostsList/Entry.jsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types"
 import { Link } from "gatsby"
 import dateFormat from "dateformat"
 
-import styles from "./entry.module.scss"
+import styles from "./Entry.module.scss"
 
 const renderAuthor = ({ className, key, name }) => {
   if (!key || !name) return null
@@ -43,7 +43,7 @@ Entry.propTypes = {
     name: PropTypes.string.isRequired,
   }),
   date: PropTypes.instanceOf(Date).isRequired,
-  intro: PropTypes.string.isRequired,
+  intro: PropTypes.string,
   path: PropTypes.string.isRequired,
   title: PropTypes.string.isRequired,
 }
