+++ src/pages/blog.js
@@ -3,7 +3,7 @@ import { useStaticQuery, graphql } from "gatsby"
 
 import Layout from "~/src/components/layout"
 import SEO from "~/src/components/SEO"
-import PostsList from "~/src/components/blog/posts_list"
+import PostsList from "~/src/components/blog/PostsList"
 
 import "../common/base.scss"
 import styles from "./blog.module.scss"
